% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_network.R
\name{dynamic_network}
\alias{dynamic_network}
\title{Dynamic network}
\usage{
dynamic_network(DT = NULL, id = NULL, by = NULL)
}
\arguments{
\item{DT}{`data.table`` of relocations.}

\item{id}{individual identifier column name.}

\item{by}{columns in input DT to split home range network generation and comparison by. For example: c('season', 'year') or 'herd'. Expects character vector.}
}
\value{
Graph strength for each individual.
}
\description{
Dynamic network
}
