% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_network.R
\name{hr_network}
\alias{hr_network}
\title{Homerange Networks}
\usage{
hr_network(DT = NULL, id = NULL, utm = NULL, by = NULL,
  returns = NULL)
}
\arguments{
\item{DT}{`data.table`` of relocations.}

\item{id}{individual identifier column name.}

\item{utm}{proj4string indicating coordinate system of coordinates}

\item{by}{columns in input DT to split home range network generation and comparison by. For example: c('season', 'year') or 'herd'. Expects character vector.}

\item{returns}{either 'network-stats' or 'overlap'. See Details.}
}
\value{
graph strength for each individual
}
\description{
Build home range networks using \code{adehabitatHR::kerneloverlap} and returns either graph statitics or home range overlap.
}
\details{
\code{DT} provided with columns EASTING, NORTHING for UTM coordinates. \code{by} argument used to specify grouping. Defaults only (as used in the paper).
}
