% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_length.R
\name{step_length}
\alias{step_length}
\title{Step Length}
\usage{
step_length(DT, coords = c("EASTING", "NORTHING"), time = "datetime",
  splitBy = c("id", "yr"), moverate = FALSE, type = "lag",
  preserve = FALSE)
}
\arguments{
\item{DT}{data.table}

\item{coords}{character vector, length 2, coordinate column names. UTM required.}

\item{time}{character time column name.}

\item{splitBy}{character vector of column names to split step length calculation by. default is id and yr (individual identifier and year as numeric).}

\item{moverate}{calculate movement rate? stepLength / dif time, unit hours.}

\item{type}{default: lag. alternative: lead.}

\item{preserve}{preserve intermediate cols? default: no.}
}
\description{
Calculate basic step length with data.table
}
\examples{

# Load data.table
library(data.table)

# Read example data
DT <- fread(system.file("extdata", "DT.csv", package = "toast"))

DT[, datetime := as.POSIXct(datetime)]

DT[, yr := year(datetime)]

step_length(DT, coords = c('X', 'Y'), splitBy = c('ID', 'yr'))

}
